requirements.txt
asn1crypto==0.24.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
coverage==4.5.1
coveralls==1.5.0
cryptography==2.3.1
dj-database-url==0.5.0
Django==2.1.2
django-cors-headers==2.4.0
django-extensions==2.1.2
django-heroku==0.3.1
django-nose==1.4.5
djangorestframework==3.8.2
docopt==0.6.2
gunicorn==19.9.0
idna==2.7
nose==1.3.7
psycopg2==2.7.5
pycparser==2.19
PyJWT==1.6.4
python-dotenv==0.9.1
python-http-client==3.1.0
pytz==2018.5
requests==2.19.1
sendgrid==5.6.0
six==1.11.0
urllib3==1.23
whitenoise==4.1
sendgrid==5.6.0
autopep8==1.4
pep8==1.7.1